# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Public Benefits Navigator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfile {
  id: number;
  name: string;
  state: string;
  age: number;
  /** @nullable */
  income?: number | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  createdAt: string | null;
}

export interface UserProfileInput {
  name?: string;
  state?: string;
  age?: number;
  /** @nullable */
  income?: number | null;
  /** @nullable */
  occupation?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  category?: string | null;
}

export type SchemeSchemeType = typeof SchemeSchemeType[keyof typeof SchemeSchemeType];


export const SchemeSchemeType = {
  central: 'central',
  state: 'state',
} as const;

export interface Faq {
  question: string;
  answer: string;
}

export interface Scheme {
  id: string;
  title: string;
  category: string;
  description: string;
  benefits: string;
  eligibility: string[];
  documents: string[];
  applicationProcess: string[];
  officialWebsite: string;
  schemeType: SchemeSchemeType;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  minAge?: number | null;
  /** @nullable */
  maxAge?: number | null;
  /** @nullable */
  maxIncome?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  occupation?: string | null;
  isFeatured?: boolean;
  /** URL of a representative image for the scheme */
  image?: string;
  faqs?: Faq[];
  lastUpdated: string;
}

export interface SchemesListResponse {
  schemes: Scheme[];
  total: number;
  page: number;
  limit: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface SchemeStats {
  totalSchemes: number;
  categoryCounts: CategoryCount[];
  centralSchemes: number;
  stateSchemes: number;
  totalBeneficiaries: string;
}

export interface SavedScheme {
  id: number;
  schemeId: string;
  scheme: Scheme;
  /** @nullable */
  notes?: string | null;
  savedAt: string;
}

export interface SaveSchemeInput {
  /** @nullable */
  notes?: string | null;
}

export interface EligibilityInput {
  age: number;
  state: string;
  income: number;
  gender: string;
  occupation: string;
  category: string;
}

export interface SchemeResult {
  scheme: Scheme;
  reason: string;
  matchScore: number;
  requiredDocuments: string[];
  /** @nullable */
  missingCriteria?: string | null;
}

export interface EligibilityResponse {
  eligible: SchemeResult[];
  partiallyEligible: SchemeResult[];
  totalEstimatedBenefit: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  messages: Message[];
  createdAt: string;
  updatedAt: string;
}

export interface ChatMessageInput {
  content: string;
}

export interface ChatMessageResponse {
  userMessage: Message;
  assistantMessage: Message;
}

export type ListSchemesParams = {
search?: string;
category?: string;
state?: string;
schemeType?: ListSchemesSchemeType;
gender?: string;
minAge?: number;
maxAge?: number;
maxIncome?: number;
occupation?: string;
page?: number;
limit?: number;
};

export type ListSchemesSchemeType = typeof ListSchemesSchemeType[keyof typeof ListSchemesSchemeType];


export const ListSchemesSchemeType = {
  central: 'central',
  state: 'state',
  all: 'all',
} as const;