# lib/api-zod/src/generated/types/eligibilityResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Public Benefits Navigator API
 * OpenAPI spec version: 0.1.0
 */
import type { SchemeResult } from './schemeResult';

export interface EligibilityResponse {
  eligible: SchemeResult[];
  partiallyEligible: SchemeResult[];
  totalEstimatedBenefit: string;
}