# lib/api-zod/src/generated/types/saveSchemeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Public Benefits Navigator API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveSchemeInput {
  /** @nullable */
  notes?: string | null;
}