# lib/api-zod/src/generated/types/savedScheme.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Public Benefits Navigator API
 * OpenAPI spec version: 0.1.0
 */
import type { Scheme } from './scheme';

export interface SavedScheme {
  id: number;
  schemeId: string;
  scheme: Scheme;
  /** @nullable */
  notes?: string | null;
  savedAt: string;
}