# lib/api-zod/src/generated/types/scheme.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Public Benefits Navigator API
 * OpenAPI spec version: 0.1.0
 */
import type { Faq } from './faq';
import type { SchemeSchemeType } from './schemeSchemeType';

export interface Scheme {
  id: string;
  title: string;
  category: string;
  description: string;
  benefits: string;
  eligibility: string[];
  documents: string[];
  applicationProcess: string[];
  officialWebsite: string;
  schemeType: SchemeSchemeType;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  minAge?: number | null;
  /** @nullable */
  maxAge?: number | null;
  /** @nullable */
  maxIncome?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  occupation?: string | null;
  isFeatured?: boolean;
  /** URL of a representative image for the scheme */
  image?: string;
  faqs?: Faq[];
  lastUpdated: string;
}