# lib/api-zod/src/generated/types/schemeResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Public Benefits Navigator API
 * OpenAPI spec version: 0.1.0
 */
import type { Scheme } from './scheme';

export interface SchemeResult {
  scheme: Scheme;
  reason: string;
  matchScore: number;
  requiredDocuments: string[];
  /** @nullable */
  missingCriteria?: string | null;
}